Move memstore into its own package, eliminate the need for a separate protocol_test package.
