+++ core/account/reserve_test.go
@@ -8,6 +8,7 @@ import (
 	"chain/database/pg/pgtest"
 	"chain/protocol/bc"
 	"chain/protocol/prottest"
+	"chain/protocol/prottest/memstore"
 )
 
 const sampleAccountUTXOs = `
@@ -36,7 +37,7 @@ func TestCancelReservation(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	c := prottest.NewChainWithStorage(t, prottest.NewMemStore(), outid)
+	c := prottest.NewChainWithStorage(t, memstore.New(), outid)
 
 	utxoDB := newReserver(db, c, nil)
 	res, err := utxoDB.ReserveUTXO(ctx, outid, nil, time.Now())
