+++ core/txbuilder/finalize_test.go
@@ -299,7 +299,7 @@ func benchGenBlock(b *testing.B) {
 	b.StopTimer()
 
 	ctx := context.Background()
-	c := prottest.NewChainWithStorage(b, prottest.NewMemStore())
+	c := prottest.NewChainWithStorage(b, memstore.New())
 	g := generator.New(c, nil, pgtest.NewTx(b))
 	initialBlock, err := c.GetBlock(ctx, 1)
 	if err != nil {
