+++ protocol/block_test.go
@@ -1,4 +1,4 @@
-package protocol_test
+package protocol
 
 import (
 	"context"
@@ -6,9 +6,8 @@ import (
 	"testing"
 	"time"
 
-	. "chain/protocol"
 	"chain/protocol/bc"
-	"chain/protocol/prottest"
+	"chain/protocol/prottest/memstore"
 	"chain/protocol/state"
 	"chain/testutil"
 )
@@ -17,8 +16,8 @@ func TestGetBlock(t *testing.T) {
 	ctx := context.Background()
 
 	b1 := &bc.Block{BlockHeader: bc.BlockHeader{Height: 1}}
-	noBlocks := prottest.NewMemStore()
-	oneBlock := prottest.NewMemStore()
+	noBlocks := memstore.New()
+	oneBlock := memstore.New()
 	oneBlock.SaveBlock(ctx, b1)
 	oneBlock.SaveSnapshot(ctx, 1, state.Empty())
 
@@ -48,17 +47,17 @@ func TestGetBlock(t *testing.T) {
 
 func TestNoTimeTravel(t *testing.T) {
 	ctx := context.Background()
-	c, err := NewChain(ctx, bc.Hash{}, prottest.NewMemStore(), nil)
+	c, err := NewChain(ctx, bc.Hash{}, memstore.New(), nil)
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	SetChainHeight(c, 1)
-	SetChainHeight(c, 2)
+	c.setHeight(1)
+	c.setHeight(2)
 
-	SetChainHeight(c, 1) // don't go backward
-	if c.Height() != 2 {
-		t.Fatalf("c.state.height = %d want 2", c.Height())
+	c.setHeight(1) // don't go backward
+	if c.state.height != 2 {
+		t.Fatalf("c.state.height = %d want 2", c.state.height)
 	}
 }
 
@@ -201,7 +200,7 @@ func TestValidateBlockForSig(t *testing.T) {
 	}
 
 	ctx := context.Background()
-	c, err := NewChain(ctx, initialBlock.Hash(), prottest.NewMemStore(), nil)
+	c, err := NewChain(ctx, initialBlock.Hash(), memstore.New(), nil)
 	if err != nil {
 		t.Fatal("unexpected error ", err)
 	}
@@ -224,7 +223,7 @@ func newTestChain(tb testing.TB, ts time.Time) (c *Chain, b1 *bc.Block) {
 	if err != nil {
 		testutil.FatalErr(tb, err)
 	}
-	c, err = NewChain(ctx, b1.Hash(), prottest.NewMemStore(), nil)
+	c, err = NewChain(ctx, b1.Hash(), memstore.New(), nil)
 	if err != nil {
 		testutil.FatalErr(tb, err)
 	}
