--- protocol/export_test.go
@@ -1,5 +1,0 @@
-package protocol
-
-func SetChainHeight(c *Chain, height uint64) {
-	c.setHeight(height)
-}
